refactor: big refactor to split the webpack init.ts file away from the core logic to initialise and bootstrap Sentry if the user logs an error and is opted in for logging
